feat(util): allow custom attributes on components
Users desire to use attributes such as `data-hook` on our components; so,
starting with a somewhat small list of desired attributes, filter the props and
apply them to the top-level element in the component. This change starts with
LargeCardKeyValue, and will need to be applied to the other components.

BREAKING CHANGE: In LargeCardKeyValue, the `name` attribute has been changed to
`label`, so that the top-level element in the component can support an
arbitrary `name` attribute, and not conflict.