fix(BarChart): relative icon positioning
If the icon fits within the filled-in rectangle of the BarChart, then display
it at a fixed vertical position (same behavior as previous commit but a little
lower on the y-axis). Otherwise, the icon floats a few pixels above the height
of the filled-in rectangle (new behavior from this commit).

Renamed the cx/cy attributes to more correctly identify them as tooltip
positions, since cx/cy are legitimate SVG attributes but not for rectangles,
which was a little confusing to figure out.

Moved the tooltip mouse events to be tied to the invisible full-height
rectangle, which allows us to remove duplicated attributes, and displays the
tooltip when mousing over the expected area. This also fixes the unreported
issue where mousing over the date label would fire a tooltip that had `null` in
it.

Fixed the examples markdown to illustrate the icon behavior at different bar
heights.