fix(chart): x is really a label
Since this chart will only deal with dates for the x-axis, and they being as
labels and are used as labels, remove xDataType, and clarify the use of
`data.x` as only being a positional parameter versus both that and a label.

BREAKING CHANGE: this removes a previously required parameter, and renames
another one.