Fix micro card layout
Upgrading flexbox-react caused a number of layout issues.  The library changed how it generated styles, from one where the styles are inline on each element to one where the styles are created as classes with randomly generated names.  This works but changes the precidence of the applied styles.  Effectively the styles that were originally explcitly attached to the element, where they are explicitly the most matching are now in classes where they might be (and often are) overridden.  This isn't really a problem when you use this all of the time since you simply don't get in the position where this is a problem, but it's a problem when upgrading classes.

The fall out of this is that the Micro card is broken and these chagnes fix the issue.