fix(props) className prop types
== DETAILS

This changes the PropTypes for className to be string instead of
object, because CSS class names are strings.

== TESTING

Added a className prop to each of these components and confirmed that I
saw warnings about incorrect prop types, then confirmed that after I
changed the PropTypes, they didn't print warnings anymore.