feat(buttons): change styles to match UX preferences
== DETAILS

Our style guide has focused down on three button styles only (generic, positive and negative), and this change set just updates that subset.  In all cases, the hover and active styles are darker than the default state, the active and focus states all have outlines.

See the image for details.

Note that the focus outlines don't really play well with ButtonGroups, but we aren't using these at this point so I think that that can be deferred until a later date.  For now, there are no outlines within button groups.  Given how SUIR doesn't use outlines at all, there's not a lot of support for outlines that overlap adjacent elements.  I think that this is fixable using z-index changes in the focus style but didn't feel that it was worth the time to dig in on this for a feature that we're not using.

I've updated the Button example to remove the options that we weren't using since they were confusing developers who went and used them.

I fixed a bunch of issues where hover styles didn't apply to focused buttons - this was an old bug that was hidden by the previous similarity of the styles.  The bottom line is that there's a standard order in which to define the pseudo classes, (which is: Link, Visited, Focus, Hover, Active), and we weren't doing this, now we are.

The final twist is using a sequence of box-shadows to achieve the focus outline.  This seems to be a consistently cross-platform and cross-browser compatible way to get what we wanted, the alternatives either looked awful (IMNSHO) or were inconsistent between browsers or OSs.

== TESTS

Lots of exercising in the browser, the animation should show most of this.