feat(pagination): Add data-hooks
UI test automation has a hard time selecting the pagination control
buttons in the DOM, so this adds data-hook attributes to the DOM
elements to make them easier to find. They are not globally unique, but
another unique data-hook can be added to the parent component in order
to enable tests to find the specific buttons they need to test.