feat(nav): add NavLink support to SideNav.MenuItem
This extends MenuItem so that it can more usefully interact with
react-router NavLinks.  This means that the highlighting of the active
menu item is handled much more automatically.

Note that I did have a version of this that added automatic Popups to
manage tooltips as well, but as soon as I moved that behavior into the
library I started getting errors about accessing unmounted components.